+++ src/action/gazeArbiter/src/gazeArbiterModule.cpp
@@ -142,11 +142,11 @@ bool gazeArbiterModule::configure(yarp::os::ResourceFinder &rf) {
     /* get the dimension of the image for the thread parametric control */
     width                  = rf.check("width", 
                            Value(320), 
-                           "width of the image (int)").asInt();
+                           "width of the image (int)").asInt16();
 
     height                 = rf.check("height", 
                            Value(240), 
-                           "height of the image (int)").asInt();
+                           "height of the image (int)").asInt16();
 
     printf("\n width: %d  height:%d \n", width, height);
     arbiter->setDimension(width,height);
@@ -166,21 +166,21 @@ bool gazeArbiterModule::configure(yarp::os::ResourceFinder &rf) {
     /* offset for 3d position along x axis */
     this->xoffset       = rf.check("xoffset", 
                            Value(0), 
-                           "offset for 3D fixation point x").asDouble();
+                           "offset for 3D fixation point x").asFloat32();
     printf("xoffset:%f \n", xoffset);
     arbiter->setXOffset(xoffset);
 
     /* offset for 3d position along y axis */
     this->yoffset       = rf.check("yoffset", 
                            Value(0), 
-                           "offset for 3D fixation point y").asDouble();
+                           "offset for 3D fixation point y").asFloat32();
     printf("yoffset:%f \n", yoffset);
     arbiter->setYOffset(yoffset);
 
     /* offset for 3d position along z axis */
     this->zoffset       = rf.check("zoffset", 
                            Value(0), 
-                           "offset for 3D fixation point z").asDouble();
+                           "offset for 3D fixation point z").asFloat32();
     printf("zoffset:%f \n", zoffset);
     arbiter->setZOffset(zoffset);
 
@@ -188,40 +188,40 @@ bool gazeArbiterModule::configure(yarp::os::ResourceFinder &rf) {
     // limits for 3d position along x axis 
     xmax       = rf.check("xmax", 
                            Value(-0.2), 
-                          "limit max for 3D fixation point x").asDouble();
+                          "limit max for 3D fixation point x").asFloat32();
     printf("xmax:%f \n", xmax);
     xmin       = rf.check("xmin", 
                            Value(-10.0), 
-                           "limit min for 3D fixation point x").asDouble();;
+                           "limit min for 3D fixation point x").asFloat32();;
     printf("xmin:%f \n", xmin);
     arbiter->setXLimits(xmax,xmin);
     
     // limits for 3d position along y axis 
     ymax       = rf.check("ymax", 
                            Value(0.3), 
-                           "limit max for 3D fixation point y").asDouble();
+                           "limit max for 3D fixation point y").asFloat32();
     printf("ymax:%f \n", ymax);
     ymin       = rf.check("ymin", 
                            Value(-0.3), 
-                           "limit max for 3D fixation point y").asDouble();
+                           "limit max for 3D fixation point y").asFloat32();
     printf("ymin:%f \n", ymin);
     arbiter->setYLimits(ymax,ymin);
     
     // limits for 3d position along z axis 
     zmax       = rf.check("zmax", 
                            Value(0.9), 
-                           "limit max for 3D fixation point z").asDouble();
+                           "limit max for 3D fixation point z").asFloat32();
     printf("zmax:%f \n", zmax);
     zmin       = rf.check("zmin", 
                            Value(-0.3), 
-                           "limit min for 3D fixation point z").asDouble();
+                           "limit min for 3D fixation point z").asFloat32();
     printf("zmin:%f \n", zmin);
     arbiter->setZLimits(zmax,zmin);
     
     // specifies whether the camera is mounted on the head
     //onWings       = rf.check("onWings", 
     //                       Value(0), 
-    //                       "indicates whether the camera is mounted on the head").asInt();
+    //                       "indicates whether the camera is mounted on the head").asInt16();
     //printf("onWings %d \n", onWings);
     //arbiter->setOnWings(onWings);
     
@@ -257,7 +257,7 @@ bool gazeArbiterModule::configure(yarp::os::ResourceFinder &rf) {
     
     pitch       = rf.check("anglePitch", 
                            Value(-39), 
-                           "fixing the pitch to a desired angle").asDouble();
+                           "fixing the pitch to a desired angle").asFloat32();
     printf("pitch:%f \n", pitch);
     
     if(pitchFlag){
